requirements.txt
Flask==0.11.1
future==0.16.0
gunicorn==19.7.1
httplib2==0.9.2
line-bot-sdk==1.3.0
Janome==0.3.4
